update previous test
Alert user and make a display about the test with the name user wants
to  enter to create a new one.
transform the the quiz that is being pull from the mongoDb by splitting
its bundle into a list of quizzes-> quizs from bundletoQuizs function.
Format category just like title in order to have an id with no space.
take away obsolete hashtable and remplace with hash map improving
performance and sacrifice serialization.