Create, Read and Update complete
CRU of CRUD is complete. fix issues that would hinder record update
within database. Users can not edit straight from the table and if exam
is already in database user is notified and exam is displayed.
Another change was replacing stream() with parallelStream() in blocking
operation such as collections from a database.