Architecture change
The new architecture MOVE is adopted. MOVE: Model Operations Views
Events. In the case pf this project the Events have been replaced by
Controllers with each view is associated with a controller. A main
Operation that assembles all the common methods that all the
controllers are used and in some instances the model. The
datamanagement is the super class of all controllers since it interacts
directly with the data from databases yet it is still part of the Model
module.